% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rgGetData}
\alias{rgGetData}
\title{Retrieve data from regenesis}
\usage{
rgGetData(tabid)
}
\arguments{
\item{tabid}{tabid provided by rgSearch()}
}
\description{
\code{rgGetData()} retrieve data from regenesis site
}
\seealso{
\code{\link{regensis-package}}
}

