% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rgSearch}
\alias{rgSearch}
\title{Search regenesis for datasets}
\usage{
rgSearch(str, reg = "KREISE")
}
\arguments{
\item{str}{String to search for}

\item{reg}{Geographic attribute dafault="KREISE". see for more: http://regenesis.pudo.org/regional/index.html}
}
\description{
\code{rgSearch()} search static regensis-metadatadb shipped with regenesis
}
\seealso{
\code{\link{regensis-package}}
}

